current user set value???

+++ public/app/controllers/overviewCtrl.js
@@ -1,5 +1,4 @@
 angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', 'ngAnimate', 'ui.bootstrap'])
-        .controller('overviewController', ['$modal', 'CalcService', '$location','dateTimeSourceFormat','dateTimeTargetFormat','dateTargetFormat', function ($modal, CalcService, $location,dateTimeSourceFormat,dateTimeTargetFormat,dateTargetFormat) {
     .controller('overviewController', ['$modal', 'CalcService', '$location', 'currentUser', 'dateTimeSourceFormat', 'dateTimeTargetFormat', 'dateTargetFormat', function ($modal, CalcService, $location, currentUser, dateTimeSourceFormat, dateTimeTargetFormat, dateTargetFormat) {
         var vm = this;
         var test = 'test';
@@ -31,21 +30,26 @@ angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', '
         };
 
         vm.copyLink = function (grid, row) {
-            var linkForCopy = window.location.href.replace("overview","solve")  + '/' + row.entity._id;
+            var linkForCopy = window.location.href.replace("overview", "solve") + '/' + row.entity._id;
             vm.open('lg', linkForCopy);
         };
 
         vm.start = function (grid, row) {
-            //console.log('Start...' + 'row.entity.id: ' + row.entity.id);
-            // Todo fixe id ist noch zu ersetzen zurzeit nur für Testzweck
-           // $location.path('/solve/' + '55ee7a40d258ad641d52598e');
             $location.path('/solve/' + row.entity._id);
         };
 
         vm.delete = function (grid, row) {
-            var rowToDeleteId =row.entity._id;
-            //var rowToDeleteCreaterId = row.entity._id;
-            //if (currentUser.id !== rowToDeleteCreaterId)
+
+
+            var userId = currentUser.userId
+
+            if (currentUser.role !== "Admin") {
+                return;
+            }
+
+
+            var rowToDeleteId = row.entity._id;
+
 
             CalcService.deleteCalcSet(rowToDeleteId)
                 .success(function (data) {
@@ -74,12 +78,12 @@ angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', '
                     return true;
                 }
 
-                var momentAsString = moment(value, dateTimeSourceFormat ).format(dateTimeTargetFormat);
-                return isStringContainingTerm(momentAsString,term);
+                var momentAsString = moment(value, dateTimeSourceFormat).format(dateTimeTargetFormat);
+                return isStringContainingTerm(momentAsString, term);
             }
         }
 
-        var isStringContainingTerm = function (momentAsString,term){
+        var isStringContainingTerm = function (momentAsString, term) {
             var index = term.indexOf("\\");
             while (index >= 0) {
                 term = term.replace("\\", "");
@@ -97,15 +101,15 @@ angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', '
                 }
 
                 var momentAsString = moment(value, dateTimeSourceFormat).format(dateTargetFormat);
-                return isStringContainingTerm(momentAsString,term);
+                return isStringContainingTerm(momentAsString, term);
             }
         }
 
         vm.animationsEnabled = true;
 
         var def = [
-            {field: '_id', displayName: 'id', visible: true, enableColumnMenu: false, enableHiding: false},
-            {displayName: 'Erzeuger', field: 'creator.name',enableColumnMenu: false, enableHiding: false},
+            {displayName: 'id', field: '_id', visible: false, enableColumnMenu: false, enableHiding: false},
+            {displayName: 'Erzeuger', field: 'creator.name', enableColumnMenu: false, enableHiding: false},
             {
                 displayName: 'Erzeugungsdatum',
                 field: 'created',
@@ -114,6 +118,13 @@ angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', '
                 filter: {condition: vm.dateFilter},
                 cellFilter: 'dateTimeFormaterCreated',
                 filterCellFiltered: true
+            },
+            {
+                displayName: 'ErzeugerId',
+                field: 'creator._id',
+                visible: true,
+                enableColumnMenu: false,
+                enableHiding: false
             },
             {displayName: 'Schwierigkeitsgrad', field: 'diff_level', enableColumnMenu: false, enableHiding: false},
             {displayName: 'Meine erreichte Punktzahl', field: 'score', enableColumnMenu: false, enableHiding: false},
@@ -169,7 +180,7 @@ angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', '
         };
     }
     ])
-    .filter('dateTimeFormaterCreated', function (dateTimeSourceFormat,dateTargetFormat) {
+    .filter('dateTimeFormaterCreated', function (dateTimeSourceFormat, dateTargetFormat) {
         return function (input) {
             if (!input) {
                 return '';
@@ -181,7 +192,7 @@ angular.module('mathApp.overview', ['ngTouch', 'ui.grid', 'angular-clipboard', '
             }
         };
     })
-    .filter('dateTimeFormaterLastExec', function (dateTimeSourceFormat,dateTimeTargetFormat) {
+    .filter('dateTimeFormaterLastExec', function (dateTimeSourceFormat, dateTimeTargetFormat) {
         return function (input) {
             if (!input) {
                 return '';
